Warn if document already deleted (OCECDR-4838)

+++ Inetpub/wwwroot/cgi-bin/cdr/del-some-docs.py
@@ -103,14 +103,20 @@ class Control(Controller):
                 for id in self.ids:
                     doc = Doc(self.session, id=id)
                     cdr_id = doc.cdr_id
+                    if doc.active_status == Doc.DELETED:
+                        message = f"{cdr_id} already deleted"
+                        self.logger.warning(message)
+                        items.append(B.LI(message, B.CLASS("error")))
+                        continue
                     try:
                         doc.delete(**opts)
                         self.logger.info("Deleted %s", cdr_id)
                         message = f"{cdr_id} deleted successfully"
                         items.append(B.LI(message, B.CLASS("info")))
                         for error in doc.errors:
-                            self.logger.warning("%s: %s", cdr_id, error)
-                            items.append(B.LI(f"{cdr_id}: {error}"))
+                            message = f"{cdr_id}: {error}"
+                            self.logger.warning(message)
+                            items.append(B.LI(message, B.CLASS("error")))
                     except Exception as e:
                         self.logger.exception(cdr_id)
                         items.append(B.LI(f"{cdr_id}: {e}", B.CLASS("error")))
